% Generated by roxygen2 (4.0.1): do not edit by hand
\name{abcSimulatorWrapper}
\alias{abcSimulatorWrapper}
\title{Abc Simulator wrapper}
\usage{
abcSimulatorWrapper(distance, simulator, times)
}
\arguments{
\item{distance}{A metric function}

\item{simulator}{A model simulator}

\item{times}{A vector of observation times of the data}
}
\value{
A function for building the ABC wrapped simulators
}
\description{
Wraps up a model simulator function for use with ABC
}

