% Generated by roxygen2 (4.0.1): do not edit by hand
\name{abcSmc}
\alias{abcSmc}
\title{ABC SMC}
\usage{
abcSmc(epsilon1, simulator, times, distance, prior)
}
\arguments{
\item{epsilon1}{The initial tolerance value for the ABC SMC algorithm}

\item{simulator}{A model simulator}

\item{times}{A vector of observation times of the data}

\item{distance}{A metric function}

\item{prior}{A function which gives returns prior samples of the states and rate parameters}
}
\value{
A function which runs the ABC SMC algorithm for a desired accepted sample size, over a desired number of populations for a given adaptive tolerance selected as the alpha ntile of the distribution of distances
}
\description{
A function which returns an ABC SMC algorithm function given an initial tolerance, a
model simulator, a vector of observation times, a distance calculator function which contains
the observed data information and a prior distribution for both the rates and states.
}

