% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcmcChain}
\alias{mcmcChain}
\title{MCMC Chain}
\usage{
mcmcChain(proposalVariance, dataLikelihood, nParticles, xPrior, data, simulator)
}
\arguments{
\item{proposalVariance}{The covariance matrix to be used in the Gaussian random walk proposal}

\item{dataLikelihood}{A function which calculates the data likelihood}

\item{nParticles}{The number of particles to be used in the particle filter}

\item{xPrior}{A function which returns samples from the prior distribution on the states}

\item{data}{The observed data, as a matrix}

\item{simulator}{The model simulator to be used for advancing the state of the system given some rate parameters}
}
\value{
A fucntion whihc runs a particle MCMC chain
}
\description{
A function which does the set up work for a particle MCMC chain, returning a function which can be used to run
the chain given some initial parameter values, a number of iterations and an amount to thin samples by.
}

