% Generated by roxygen2 (4.0.1): do not edit by hand
\name{metric}
\alias{metric}
\title{Eulidean Metric}
\usage{
metric(observed)
}
\arguments{
\item{observed}{The set of observed data}
}
\value{
A function which calculates the Euclidean distance given some simulated data and a row/time index
}
\description{
Creates a function which calculates the Euclidean distances between some simulated data and an observation time index
}

