% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weightCalculator}
\alias{weightCalculator}
\title{Weight Calculator}
\usage{
weightCalculator(weights, samples, sigma)
}
\arguments{
\item{weights}{A vector of weights that will be updated}

\item{samples}{A matrix of samples that have the associated weights}

\item{sigma}{A covariance matrix according to a Gaussian proposal that has been used}
}
\value{
A function which calculates weights for a matrix of proposals
}
\description{
A function to create a weight calclulator object for use in
calculating particle weights in ABC SMC.
This example is specific to the Lotka Volterra model and priors
as a demonstration
}

